Implement model loading from W&B and refactor
